// src/app/models/notesModel.ts
export class NoteModel {
    Title: string = "";
    Description: string = "";
    Color: string = "";
    Reminder: string = "2022-05-20T11:02:38.430Z";
    ImagePaths: string = "https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885_960_720.jpg";
    IsArchive: boolean = false;
    IsPinned: boolean = false;
    IsTrash: boolean = false;
}